Moved Navbar to Header Component and merged content
